+++ openstack/autoscaling/v1/policies/results.go
@@ -4,12 +4,10 @@ import (
 	"github.com/opentelekomcloud/gophertelekomcloud"
 )
 
-// Create Result is a struct which represents the create result of policy
 type CreateResult struct {
 	golangsdk.Result
 }
 
-// Extract of CreateResult will deserialize the result of Creation
 func (r CreateResult) Extract() (string, error) {
 	var a struct {
 		ID string `json:"scaling_policy_id"`
@@ -18,12 +16,10 @@ func (r CreateResult) Extract() (string, error) {
 	return a.ID, err
 }
 
-// DeleteResult is a struct which represents the delete result.
 type DeleteResult struct {
 	golangsdk.ErrResult
 }
 
-// Policy is a struct that represents the result of get policy
 type Policy struct {
 	ID             string         `json:"scaling_group_id"`
 	Name           string         `json:"scaling_policy_name"`
@@ -49,7 +45,6 @@ type Action struct {
 	InstanceNum int    `json:"instance_number"`
 }
 
-// GetResult is a struct which represents the get result
 type GetResult struct {
 	golangsdk.Result
 }
@@ -60,12 +55,10 @@ func (r GetResult) Extract() (Policy, error) {
 	return p, err
 }
 
-// UpdateResult is a struct from which can get the result of udpate method
 type UpdateResult struct {
 	golangsdk.Result
 }
 
-// Extract will deserialize the result to group id with string
 func (r UpdateResult) Extract() (string, error) {
 	var a struct {
 		ID string `json:"scaling_policy_id"`
